requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cryptography==35.0.0
defusedxml==0.7.1
distlib==0.3.3
dj-database-url==0.5.0
dj-rest-auth==2.1.11
Django==3.2.8
django-allauth==0.45.0
django-ckeditor==6.1.0
django-crispy-forms==1.13.0
django-js-asset==1.2.2
djangorestframework==3.12.4
env==0.1.0
environs==9.3.4
filelock==3.3.2
idna==3.3
marshmallow==3.13.0
oauthlib==3.1.1
Pillow==8.3.2
platformdirs==2.4.0
psycopg2==2.9.1
pycparser==2.20
PyJWT==2.3.0
python-dotenv==0.19.1
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
virtualenv==20.10.0
whitenoise==5.3.0
